refactor(medications): extract useAddMedicationForm + step components
add-medication-wizard.tsx (1216 LOC) bundled ~25 useState calls, per-step
Zod validation, and six step components in one file. Split into:

- useAddMedicationForm hook (src/hooks/) — owns all field state, single
  onFieldChange dispatcher, bulk patch, per-step validation (SearchStep,
  ScheduleEntry, InventoryStep schemas) and reset.
- add-medication-steps/ — SearchStep, AppearanceStep, IndicationStep,
  DosageStep, ScheduleStep, InventoryStep, plus the conflict-check overlay
  extracted as a sub-component. Step components now receive
  { formState, onFieldChange, errors } rather than 10+ scalar props.

Auth gating (useAuthGate) is preserved on the search and AI-suggest paths;
no UX changes. Wizard shrank to 389 LOC.

pnpm typecheck / test / lint / build all green.
